+++ cmd/action/create/cluster/defaultcontrolplane/crs.go
@@ -29,7 +29,6 @@ func (r *runner) newCRs(releases []v1alpha1.Release, host string) (v1alpha3.Clus
 	{
 		c := v1alpha3.ClusterCRsConfig{
 			ClusterID:         r.flag.TenantCluster,
-			Credential:        key.Credential,
 			Domain:            key.DomainFromHost(host),
 			Description:       "awscnfm action create cluster onenodepool",
 			Owner:             key.Organization,
@@ -41,6 +40,10 @@ func (r *runner) newCRs(releases []v1alpha1.Release, host string) (v1alpha3.Clus
 		crs, err = v1alpha3.NewClusterCRs(c)
 		if err != nil {
 			return v1alpha3.ClusterCRs{}, microerror.Mask(err)
+		}
+
+		if key.IsOrgNamespaceVersion(c.ReleaseVersion) {
+			crs = key.MoveClusterCRsToOrgNamespace(crs, key.Organization)
 		}
 	}