+++ cmd/action/delete/cluster/runner.go
@@ -8,12 +8,12 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	pkgclient "github.com/giantswarm/awscnfm/v15/pkg/client"
 	"github.com/giantswarm/awscnfm/v15/pkg/env"
+	"github.com/giantswarm/awscnfm/v15/pkg/key"
 	"github.com/giantswarm/awscnfm/v15/pkg/label"
 )
 
@@ -59,7 +59,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		err := cpClients.CtrlClient().DeleteAllOf(
 			ctx,
 			&apiv1alpha3.Cluster{},
-			client.InNamespace(metav1.NamespaceDefault),
+			client.InNamespace(key.OrganizationNamespaceFromName(key.Organization)),
 			client.MatchingLabels{label.Cluster: r.flag.TenantCluster},
 		)
 		if errors.IsNotFound(err) {